chore: zero clippy warnings across the workspace
Deleted the dead capture_active block in jit_lower (an orphaned
brood_rt_in_capture entry-call whose result was never consumed) plus the
unused brood_rt_tick import it stranded; slot-tag scans use contains();
fn-pointer casts go through usize; the VecStore::Spill ptr/len fields get
a documented allow(dead_code) — Rust never reads them by name, JIT native
code loads them via the repr(u8)-pinned offsets. Suite 784/784; 200
differential-fuzz seeds agree post-codegen-change.

+++ crates/lisp/src/eval/compile/mod.rs
@@ -6822,7 +6822,7 @@ pub(crate) fn jit_dispatch_call(
                         // stays entirely in IR (arity pre-validated for this argc).
                         let nat = heap.native(nid);
                         if nat.arity.accepts(argc) && !value::is_dynamic(head) {
-                            let func = nat.func as u64;
+                            let func = nat.func as usize as u64;
                             heap.vm_fast_link_publish_native(site, head, argc as u32, epoch, func);
                         }
                         call_native_direct!(nid)
@@ -6871,7 +6871,7 @@ pub(crate) fn jit_dispatch_call(
                                 // straight to the fn pointer (arity pre-validated here).
                                 let nat = heap.native(nid);
                                 if nat.arity.accepts(argc) {
-                                    let func = nat.func as u64;
+                                    let func = nat.func as usize as u64;
                                     heap.vm_fast_link_publish_native(
                                         site,
                                         head,
